+++ src/init.c
@@ -8,6 +8,7 @@
 
 static const R_CallMethodDef call_methods[] = {
     {"C_read_csv", (DL_FUNC)&C_read_csv, 5},
+    {"C_sniff_csv", (DL_FUNC)&C_sniff_csv, 3},
     {NULL, NULL, 0}
 };